# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/event": {
            "post": {
                "description": "null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Add/Update event",
                "parameters": [
                    {
                        "description": "EventName",
                        "name": "eventName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/selftype.EventName"
                        }
                    },
                    {
                        "description": "[]Deagnosis",
                        "name": "deagnosis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/selftype.Deagnosis"
                            }
                        }
                    },
                    {
                        "description": "[]Procedure",
                        "name": "procedure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/selftype.Procedure"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/selftype.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    }
                }
            }
        },
        "/event/{eventName}": {
            "get": {
                "description": "Get an event from string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event search by eventName",
                        "name": "eventName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/selftype.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Delete an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event delete by eventName",
                        "name": "eventName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    }
                }
            }
        },
        "/prove": {
            "post": {
                "description": "Give a prove of ZKRP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZKRP"
                ],
                "summary": "Prove ZKRP",
                "parameters": [
                    {
                        "description": "requestTime",
                        "name": "requestTime",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "[]Deagnosis",
                        "name": "deagnosis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/selftype.DeagnosisDate"
                            }
                        }
                    },
                    {
                        "description": "[]Procedure",
                        "name": "procedure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/selftype.ProcedureDate"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/selftype.ProvePackagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    }
                }
            }
        },
        "/verify": {
            "post": {
                "description": "Verify a prove of ZKRP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZKRP"
                ],
                "summary": "Verify ZKRP",
                "parameters": [
                    {
                        "description": "[]ProvePackage",
                        "name": "provePackages",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/selftype.ProvePackage"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/selftype.JSONResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "selftype.Deagnosis": {
            "type": "object",
            "properties": {
                "deagnosisCode": {
                    "type": "string",
                    "example": "J0190"
                },
                "duringTime": {
                    "type": "integer",
                    "example": 28
                }
            }
        },
        "selftype.DeagnosisDate": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "20200505"
                },
                "deagnosisCode": {
                    "type": "string",
                    "example": "J0190"
                },
                "duringTime": {
                    "type": "integer",
                    "example": 28
                }
            }
        },
        "selftype.Event": {
            "type": "object",
            "properties": {
                "deagnosis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/selftype.Deagnosis"
                    }
                },
                "eventName": {
                    "type": "string",
                    "example": "marathon"
                },
                "procedure": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/selftype.Procedure"
                    }
                }
            }
        },
        "selftype.EventName": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "marathon"
                }
            }
        },
        "selftype.EventResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/selftype.Event"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "selftype.JSONResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "selftype.Procedure": {
            "type": "object",
            "properties": {
                "duringTime": {
                    "type": "integer",
                    "example": 28
                },
                "procedureCode": {
                    "type": "string",
                    "example": "3950"
                }
            }
        },
        "selftype.ProcedureDate": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "20200505"
                },
                "duringTime": {
                    "type": "integer",
                    "example": 28
                },
                "procedureCode": {
                    "type": "string",
                    "example": "3950"
                }
            }
        },
        "selftype.ProvePackage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "J0390"
                },
                "commitment": {
                    "type": "string",
                    "example": "0x1af2e641e834f6c503e0e6b9b593323342e7d45ad0aa299bc934110a4e13ae2f33c3a0fc695f397d51e8b4699b4d9f9397b3bc3e20fa42cb1a540ff8f19dc118"
                },
                "lowerbound": {
                    "type": "integer",
                    "example": 20200503
                },
                "prove": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "Deagnosis"
                },
                "upperbound": {
                    "type": "integer",
                    "example": 20200531
                }
            }
        },
        "selftype.ProvePackagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/selftype.ProvePackage"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "x-extension-openapi": {
        "example": "value on a json format"
    }
}